third_party/blink/web_tests/html/tabular_data/table_cellpadding-expected.txt

HTMLTableElement cellpadding attribute test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS cellPaddingAttributeEffect("") is "1px"
PASS cellPaddingAttributeEffect("1") is "1px"
PASS cellPaddingAttributeEffect("2") is "2px"
PASS cellPaddingAttributeEffect("10") is "10px"
PASS cellPaddingAttributeEffect("0") is "0px"
PASS cellPaddingAttributeEffect("-1") is "0px"
PASS cellPaddingAttributeEffect("1x") is "1px"
PASS cellPaddingAttributeEffect("1.") is "1px"
PASS cellPaddingAttributeEffect("1.9") is "1px"
PASS cellPaddingAttributeEffect("2x") is "2px"
PASS cellPaddingAttributeEffect("2.") is "2px"
PASS cellPaddingAttributeEffect("2.9") is "2px"
PASS cellPaddingAttributeEffect("a") is "0px"
PASS cellPaddingAttributeEffect(arabicIndicDigitOne) is "0px"
PASS cellPaddingAttributeEffect("2" + arabicIndicDigitOne) is "2px"
PASS successfullyParsed is true

TEST COMPLETE
